% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_table_exists.R
\name{check_if_table_exists}
\alias{check_if_table_exists}
\title{Check if a given table name exists}
\usage{
check_if_table_exists(db_con, table_name)
}
\arguments{
\item{db_con}{A DBI database connection. See `db_connection()` for details.}

\item{table_name}{(String) The name of the table or view to be checked.}
}
\value{
True, if the table exists, false otherwise.
}
\description{
See title.
}
\examples{
\dontrun{
res <- DIZutils::check_if_table_exists(
  db_con = DIZutils::db_connection(...),
  table_name = "my_table"
)}

}
