% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Clean the console and environment-variables}
\usage{
clear(keep_environment = FALSE, keep_console = FALSE)
}
\arguments{
\item{keep_environment}{(Optional, boolean) If true, the objects from the
environment will not be deleted/emptied.}

\item{keep_console}{(Optional, boolean) If true, the console will not
be emptied.}
}
\value{
Nothing.
}
\description{
Function to clean the local environment.
  The call of this function clears the console and the local environment
  variables.
}
