% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_connection.R
\name{close_connection}
\alias{close_connection}
\title{close_connection helper function}
\usage{
close_connection(conn)
}
\arguments{
\item{conn}{A DBI database connection.}
}
\value{
The function is just a wrapper around RPostgres::dbDisconnect /
  DBI::dbDisconnect and does not return any value.
}
\description{
Internal function to close the database connection. The function
  is just a wrapper around `RPostgres::dbDisconnect`
}
\examples{
\dontrun{
db_con <- DIZutils::db_connection(
  db_name = "i2b2",
  db_type = "postgres"
)

DIZutils::close_connection(
  conn = db_con
)
}

}
