% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstup.R
\name{firstup}
\alias{firstup}
\title{Converts the first letter of the input string to uppercase}
\usage{
firstup(x)
}
\arguments{
\item{x}{A character string. E.g. "hello world" will become "Hello world".}
}
\value{
Returns the input string but with a capital first letter.
}
\description{
Converts the first letter of the input string to uppercase
}
\examples{
firstup("first letter of this string will be upper case as return")

}
