% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_systems.R
\name{get_db_systems}
\alias{get_db_systems}
\title{Quickly get all currently implemented database systems}
\usage{
get_db_systems()
}
\value{
The currently implemented database systems as string array.
  `E.g. c("postgres", "oracle")`
#'
}
\description{
Function to quickly get the currently implemented database
  systems
}
\examples{
get_db_systems()
# Result: c("postgres", "oracle")

}
