% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_env.R
\name{get_default_config_elements}
\alias{get_default_config_elements}
\title{Get default element names (suffixes) for env reading}
\usage{
get_default_config_elements()
}
\value{
Vector with default elements
}
\description{
Internal function.
  Get default element names (suffixes) for env reading
}
