% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_env_hack.R
\name{global_env_hack}
\alias{global_env_hack}
\title{global_env_hack}
\usage{
global_env_hack(key, val, pos = 1)
}
\arguments{
\item{key}{A character (!) string. The name of the assigned variable}

\item{val}{An object. The object that will be assigned to 'key'.}

\item{pos}{An integer. The position of the environment (default: 1).}
}
\value{
No return value, called for side effects (see description).
}
\description{
Hack variable into global env (bypasses R CMD checks).
  This does create a new variable in the R environment but NOT a new
  variable in the system environment. To create a system environment
  variable beeing accessible via `Sys.getenv(...)`, use the function
  `DIZutils::setenv2(key = "varname", val = 7)`.
}
\examples{
utils_path <- tempdir()
global_env_hack(
  key = "utils_path",
  val = utils_path,
  pos = 1L
)

}
\seealso{
\url{http://adv-r.had.co.nz/Environments.html}
}
