% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%not in\%}
\alias{\%not in\%}
\title{notin helper function}
\usage{
x \%not in\% y
}
\arguments{
\item{x}{Object 1.}

\item{y}{Object 2.}
}
\description{
Function to return elements of x that are not in y.
}
