% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{notin helper function}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{Object 1.}

\item{y}{Object 2.}
}
\value{
Returns the result of !%in%(x,y)
}
\description{
Function to return elements of x that are not in y.
}
\examples{
tmp1 <- c("a","b","c")
tmp2 <- c("b", "c", "d")
tmp1 \%notin\% tmp2
}
