% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date_format.R
\name{is_date_format}
\alias{is_date_format}
\title{Checks if a string matches a given date format.}
\usage{
is_date_format(date, format)
}
\arguments{
\item{date}{The list applied from rv$restricting_date}

\item{format}{The format parameters. See \code{?strptime}
for parameter info.}
}
\value{
TRUE/FALSE
}
\description{
Checks if a string matches a given date format.
}
