% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_database.R
\name{query_database}
\alias{query_database}
\title{query_database helper function}
\usage{
query_database(db_con, sql_statement)
}
\arguments{
\item{db_con}{A DBI database connection.}

\item{sql_statement}{A character string containing a valid SQL statement.}
}
\value{
Returns the result of the db-query.
}
\description{
Internal function to query the database. The function
  sends a sql statement to the database and returns a data.table.
}
\examples{
\dontrun{
db_con <- DIZutils::db_connection(
  db_name = "i2b2",
  db_type = "postgres"
)

query_database(
  db_con = db_con,
  sql_statement = "SELECT * FROM table_name;"
)}

}
