% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env_vars.R
\name{set_env_vars}
\alias{set_env_vars}
\title{set_env_vars helper function}
\usage{
set_env_vars(env_file)
}
\arguments{
\item{env_file}{A character. The full path including the file name to
the file containing the environment variable defintions to be loaded.}
}
\description{
Internal function to set environment variables that are
  necessary for the database connections with \code{db_connection}.
}
\seealso{
\code{Sys.setenv}
}
