% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_2_json.R
\name{xml_2_json}
\alias{xml_2_json}
\title{Quickly transform a xml objet into a json object.}
\usage{
xml_2_json(xml)
}
\arguments{
\item{xml}{An xml object.}
}
\value{
The json-representation of the xml object.
}
\description{
See title.
}
