% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{db_connection}
\alias{db_connection}
\title{db_connection helper function}
\usage{
db_connection(db_name, headless = FALSE, timeout = 30, logfile_dir = NULL)
}
\arguments{
\item{db_name}{A character. Name of the database system. Currently
implemented systems are: 'postgres', 'oracle'.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}

\item{timeout}{A timeout in sec. for the db-connection establishment.
Values below 2 seconds are not recommended.
Default is 30 seconds.}
}
\description{
Internal function to test and get the database connection of
  the target data system.
}
