% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{db_connection}
\alias{db_connection}
\title{db_connection helper function}
\usage{
db_connection(
  db_name,
  db_type,
  headless = FALSE,
  from_env = TRUE,
  settings = NULL,
  timeout = 30,
  logfile_dir = NULL,
  lib_path = NULL
)
}
\arguments{
\item{db_name}{A character. Name of the database system.}

\item{db_type}{A character. Type of the database system. Currently
implemented systems are: 'postgres', 'oracle'.}

\item{headless}{A boolean (default: FALSE). Indicating, if the function is
run only in the console (headless = TRUE) or on a GUI frontend
(headless = FALSE).}

\item{from_env}{A boolean (default: TRUE). Should database connection
be read from the environment or from a settings file.}

\item{settings}{A list. Required if `from_env=TRUE`. A list containing
settings for the database connection. Required fields are 'host',
'db_name', 'port', 'user' and 'password'.
Additionally for Oracle DB's: 'sid'.}

\item{timeout}{A timeout in sec. for the db-connection establishment.
Values below 2 seconds are not recommended.
Default is 30 seconds.}

\item{logfile_dir}{(Optional, String, default: "tempdir()")
The absolute path to folder where the logfile will be stored.}

\item{lib_path}{A character string. The path to the ojdbc7.jar file.}
}
\value{
If successful, the result will be the established connection.
  Otherwise the result will be null.
}
\description{
Internal function to test and get the database connection of
  the target data system.
}
\examples{
db_con <- DIZutils::db_connection(
  db_name = "i2b2",
  db_type = "postgres",
  headless = true,
  logfile_dir = "./path/to/logfile/dir/"
)

}
